Instsantiate generic API tests for SQL storage

+++ sequencer/api/migrations/V11__timestamp_index.sql
@@ -1,1 +1,1 @@
-CREATE INDEX header_timesteamp ON header (data->'timestamp');
+CREATE INDEX header_timesteamp ON header (((data->'timestamp')::bigint));
